move crafting.ts and create crafting/index
