+++ src/lib/skilling/skills/crafting/index.ts
@@ -1,0 +1,26 @@
+import { Craftable } from '../../types';
+import Birdhouse from './birdhouse';
+import Built from './built';
+import Dragonhide from './dragonhide';
+import Gems from './gems';
+import Glassblowing from './glassblowing';
+import Gold from './gold';
+import Leather from './leather';
+import Misc from './misc';
+import Silver from './silver';
+import Tanning from './tanning';
+
+const craftables: Craftable[] = [
+	...Birdhouse,
+	...Built,
+	...Dragonhide,
+	...Gems,
+	...Glassblowing,
+	...Gold,
+	...Leather,
+	...Misc,
+	...Silver,
+	...Tanning
+];
+
+export default craftables;
